+++ packages/server/lib/controllers/v1/invite/deleteInvite.integration.test.ts
@@ -1,9 +1,9 @@
 import db, { multipleMigrations } from '@nangohq/database';
 import { inviteEmail, seeders } from '@nangohq/shared';
 import { afterAll, beforeAll, describe, expect, it } from 'vitest';
-import { isSuccess, runServer, shouldBeProtected, shouldRequireQueryEnv } from '../../../../utils/tests.js';
+import { isSuccess, runServer, shouldBeProtected, shouldRequireQueryEnv } from '../../../utils/tests.js';
 
-const route = '/api/v1/team/invite';
+const route = '/api/v1/invite';
 let api: Awaited<ReturnType<typeof runServer>>;
 describe(`DELETE ${route}`, () => {
     beforeAll(async () => {
