+++ packages/server/lib/controllers/v1/invite/postInvite.integration.test.ts
@@ -1,9 +1,9 @@
 import { multipleMigrations } from '@nangohq/database';
 import { seeders } from '@nangohq/shared';
 import { afterAll, beforeAll, describe, expect, it } from 'vitest';
-import { runServer, shouldBeProtected, shouldRequireQueryEnv } from '../../../../utils/tests.js';
+import { runServer, shouldBeProtected, shouldRequireQueryEnv } from '../../../utils/tests.js';
 
-const route = '/api/v1/team/invite';
+const route = '/api/v1/invite';
 let api: Awaited<ReturnType<typeof runServer>>;
 describe(`POST ${route}`, () => {
     beforeAll(async () => {
