+++ packages/server/lib/controllers/v1/invite/postInvite.ts
@@ -1,9 +1,9 @@
-import { asyncWrapper } from '../../../../utils/asyncWrapper.js';
+import { asyncWrapper } from '../../../utils/asyncWrapper.js';
 import { requireEmptyQuery, zodErrorToHTTP } from '@nangohq/utils';
 import type { PostInvite } from '@nangohq/types';
 import { expirePreviousInvitations, inviteEmail, userService } from '@nangohq/shared';
 import { z } from 'zod';
-import { sendInviteEmail } from '../../../../helpers/email.js';
+import { sendInviteEmail } from '../../../helpers/email.js';
 import db from '@nangohq/database';
 
 const validation = z
