+++ packages/server/lib/routes.ts
@@ -59,8 +59,8 @@ import { getTeam } from './controllers/v1/team/getTeam.js';
 import { putTeam } from './controllers/v1/team/putTeam.js';
 import { putResetPassword } from './controllers/v1/account/putResetPassword.js';
 import { postForgotPassword } from './controllers/v1/account/postForgotPassword.js';
-import { postInvite } from './controllers/v1/team/invite/postInvite.js';
-import { deleteInvite } from './controllers/v1/team/invite/deleteInvite.js';
+import { postInvite } from './controllers/v1/invite/postInvite.js';
+import { deleteInvite } from './controllers/v1/invite/deleteInvite.js';
 import { deleteTeamUser } from './controllers/v1/team/users/deleteTeamUser.js';
 
 export const router = express.Router();
@@ -179,8 +179,8 @@ web.route('/api/v1/meta').get(webAuth, environmentController.meta.bind(environme
 web.route('/api/v1/team').get(webAuth, getTeam);
 web.route('/api/v1/team').put(webAuth, putTeam);
 web.route('/api/v1/team/users/:id').delete(webAuth, deleteTeamUser);
-web.route('/api/v1/team/invite').post(webAuth, postInvite);
-web.route('/api/v1/team/invite').delete(webAuth, deleteInvite);
+web.route('/api/v1/invite').post(webAuth, postInvite);
+web.route('/api/v1/invite').delete(webAuth, deleteInvite);
 web.route('/api/v1/account/admin/switch').post(webAuth, accountController.switchAccount.bind(accountController));
 
 web.route('/api/v1/environment').get(webAuth, environmentController.getEnvironment.bind(environmentController));
