+++ packages/types/lib/invitations/api.ts
@@ -2,7 +2,7 @@ import type { Endpoint } from '../api';
 
 export type PostInvite = Endpoint<{
     Method: 'POST';
-    Path: '/api/v1/team/invite';
+    Path: '/api/v1/invite';
     Querystring: { env: string };
     Body: { emails: string[] };
     Success: {
@@ -12,7 +12,7 @@ export type PostInvite = Endpoint<{
 
 export type DeleteInvite = Endpoint<{
     Method: 'DELETE';
-    Path: '/api/v1/team/invite';
+    Path: '/api/v1/invite';
     Querystring: { env: string };
     Body: { email: string };
     Success: {
